\name{seqData}
\alias{seqData}
\docType{data}
\title{
  A \code{ChIPComp} object.
}
\description{
  The object is sampled from 50 common binding sites between Helas3 and K562 
  cell lines for H3K27ac and 5 unique binding sites for each cell line.
}


\usage{
	data(seqData)
}

\value{
	A "ChIPComp" class object
}

\examples{
	data(seqData)
}
\keyword{datasets}


